Add files via upload
Bob's "Down in Front" Aggravation Metric

Nothing bothers Bob more than when taller people stand up in front and block his view at sporting events. It bothers him so much, that he has developed a measure for the displeasure that a crowd feels because of this phenomenon. Most days, he calls his measure the Index of Identifiable Inversions. Although when his view is blocked by this phenomenon, he calls it the "Down In Front" Aggravation Metric.

The measure works as follows. In a line of standing people, the measure is equal to the number of people who are blocking others' views. Person A is blocking B's view, if

A is taller than B

A is standing in front of B

There is no one of equal height or taller than A in line between A and B

Assuming that everyone is standing, the line below has a "Down in Front" Aggravation Measure of 5. Bob's view is blocked by Alice, Chuck's view is blocked by Alice, Dave's view is blocked by both Chuck and Alice, and Eve's view is blocked by Alice.